% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-module.R
\name{gen_server_reactive}
\alias{gen_server_reactive}
\title{generate a server reactive}
\usage{
gen_server_reactive(.input, .input_name)
}
\arguments{
\item{.input}{input list}

\item{.input_name}{name for input}
}
\description{
generate a server reactive
}
