% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-module.R
\name{gen_ui_input}
\alias{gen_ui_input}
\title{generate an input}
\usage{
gen_ui_input(.input, .input_name)
}
\arguments{
\item{.input}{input list}

\item{.input_name}{name for input}
}
\description{
generate an input
}
